+++ app/lib/constants/api/api_const.dart
@@ -9,7 +9,7 @@ class ApiConst {
 
   static const domain = 'https://myquran.life';
 
-  String get socketBase => 'wss://$_getDomain/ws';
+  String get socketBase => 'wss://myquran.life/ws';
   String get loginWithGoogle => '$_getDomain/api/v1/accounts/google/';
   String get loginWithApple => '$_getDomain/api/v1/accounts/apple/';
   String putProfile(String userId) => '$_getDomain/api/v1/accounts/profile/$userId/';
