+++ app/lib/modules/hatim/presentation/presentation.dart
@@ -1,6 +1,7 @@
 export 'widgets/hatim_page_status_card.dart';
 export 'widgets/juz_percent_widget.dart';
 export 'widgets/vertical_text.dart';
+export 'widgets/hatim_juz_list_builder.dart';
 export 'view/hatim_select_page_view.dart';
 export 'view/hatim_view.dart';
 export 'bloc/hatim_bloc.dart';
